+++ lib/domain/record/components/supports/components/display_number_setting/store.dart
@@ -2,7 +2,7 @@ import 'package:pilll/database/batch.dart';
 import 'package:pilll/domain/record/record_page_store.dart';
 import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
-import 'package:pilll/service/pill_sheet_modified_history.dart';
+import 'package:pilll/database/pill_sheet_modified_history.dart';
 
 import 'state.codegen.dart';
 import 'package:riverpod/riverpod.dart';
