+++ lib/native/pill.dart
@@ -7,7 +7,7 @@ import 'package:pilll/database/database.dart';
 import 'package:pilll/domain/record/util/take.dart';
 import 'package:pilll/service/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
-import 'package:pilll/service/pill_sheet_modified_history.dart';
+import 'package:pilll/database/pill_sheet_modified_history.dart';
 import 'package:pilll/util/datetime/day.dart';
 
 Future<void> recordPill() async {
