+++ test/domain/settings/components/today_pill_number/setting_today_pill_number_store_test.dart
@@ -10,7 +10,7 @@ import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/service/day.dart';
 import 'package:pilll/service/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
-import 'package:pilll/service/pill_sheet_modified_history.dart';
+import 'package:pilll/database/pill_sheet_modified_history.dart';
 import 'package:pilll/util/datetime/day.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
